[Demo]: added new 'src/ts/demo/' subdir.
Changelog:
- added demo app theming;
- moved 'app.ts' and 'dataset.ts' to 'demo/' subdir.
